chore: nits and code shuffling
