+++ Cargo.toml
@@ -41,30 +41,30 @@ std = [
     "winter_crypto/std",
     "winter_math/std",
     "winter_utils/std",
+    "rand/std",
 ]
 
 [dependencies]
 blake3 = { version = "1.5", default-features = false }
 clap = { version = "4.5", features = ["derive"], optional = true }
-getrandom = { version = "0.2", features = ["js"] }
-num = { version = "0.4.1", default-features = false, features = ["alloc", "rand"] }
+num = { version = "0.4", default-features = false, features = ["alloc", "rand"] }
 num-complex = { version = "0.4.4", default-features = false }
-rand = { version = "0.8.5", default-features = false, features = ["getrandom"] }
+rand = { version = "0.8", default-features = false, features = ["alloc"] }
 rand_utils = { version = "0.8", package = "winter-rand-utils", optional = true }
 serde = { version = "1.0", features = [
     "derive",
 ], default-features = false, optional = true }
-sha3 = { version = "0.10.8", default-features = false }
+sha3 = { version = "0.10", default-features = false }
 winter_crypto = { version = "0.8", package = "winter-crypto", default-features = false }
 winter_math = { version = "0.8", package = "winter-math", default-features = false }
 winter_utils = { version = "0.8", package = "winter-utils", default-features = false }
 
 [dev-dependencies]
-seq-macro = { version = "0.3" }
 criterion = { version = "0.5", features = ["html_reports"] }
+hex = { version = "0.4", default-features = false, features = ["alloc"] }
 proptest = "1.4"
 rand_utils = { version = "0.8", package = "winter-rand-utils" }
-hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
+seq-macro = { version = "0.3" }
 
 [build-dependencies]
 cc = { version = "1.0", features = ["parallel"], optional = true }
